requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.0
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai==0.28.1
openai-whisper==20230918
packaging==23.1
pandas==2.1.1
patsy==0.5.3
Pillow==10.0.1
PyAudio==0.2.13
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
